# Frontend/openapi/backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MonDexSharp.Backend | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/species": {
      "post": {
        "tags": [
          "PokemonSpecies"
        ],
        "operationId": "CreatePokemonSpecies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonSpeciesDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonSpeciesDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonSpeciesDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "PokemonSpecies"
        ],
        "operationId": "GetPokemonSpecies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonSpeciesDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonSpeciesDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonSpeciesDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/species/{id}": {
      "get": {
        "tags": [
          "PokemonSpecies"
        ],
        "operationId": "GetPokemonSpeciesById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PokemonSpecies"
        ],
        "operationId": "UpdatePokemonSpecies",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonSpeciesDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonSpeciesDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonSpeciesDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonSpeciesDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PokemonSpecies"
        ],
        "operationId": "DeletePokemonSpecies",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/types": {
      "post": {
        "tags": [
          "PokemonTypes"
        ],
        "operationId": "CreatePokemonType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPokemonTypeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonTypeDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "PokemonTypes"
        ],
        "operationId": "GetPokemonTypes",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonTypeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonTypeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PokemonTypeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/types/{id}": {
      "get": {
        "tags": [
          "PokemonTypes"
        ],
        "operationId": "GetPokemonType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokemonTypeDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PokemonSpeciesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "genera": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "base_stats": {
            "$ref": "#/components/schemas/PokemonStatsDto"
          }
        }
      },
      "PokemonStatsDto": {
        "type": "object",
        "properties": {
          "hp": {
            "type": "integer",
            "format": "int32"
          },
          "attack": {
            "type": "integer",
            "format": "int32"
          },
          "defense": {
            "type": "integer",
            "format": "int32"
          },
          "special_attack": {
            "type": "integer",
            "format": "int32"
          },
          "special_defense": {
            "type": "integer",
            "format": "int32"
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PokemonTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpsertPokemonSpeciesDto": {
        "required": [
          "number",
          "name",
          "genera",
          "description",
          "types",
          "base_stats"
        ],
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "genera": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "base_stats": {
            "$ref": "#/components/schemas/PokemonStatsDto"
          }
        }
      },
      "UpsertPokemonTypeDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "PokemonSpecies"
    },
    {
      "name": "PokemonTypes"
    }
  ]
}